pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "langgraph-agent"
version = "0.1.0"
description = "My LangGraph agent project"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.121.3",
    "google-genai>=1.47.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-core>=0.3.80",
    "langgraph>=0.6.11",
    "pydantic>=2.12.4",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
"*" = ["*.py"]
